chore(logout-cache): reset cache when user log out
